ENH: Allow rolling multiple axes at the same time.
A quick test suggests that this implementation from @seberg, relying on
slices rather than index arrays, is 1.5~3x faster than the previous (1D)
roll (depending on the axis).

Also switched the error message for invalid inputs to match the one of
ufuncs, because the axis can actually also be negative.