BUG: allow graceful recovery for no compiler
If there is no compiler on Linux, the error we were getting was CompileError
rather than the OSError the test was expecting.

This had the nasty side-effect of leaving us in a deleted temporary directory,
causing later test failures.

Try a check to see if we have a compiler and skip otherwise.