ENH: Make assert_almost_equal & assert_array_almost_equal consistent.
This changes the check for scalars in assert_almost_equal so that

    abs(actual - desired) < 1.5 * 10**(-decimal)

Note that the previous documentation claimed that the functions were
equivalent to

    abs(actual - desired) < .5 * 10**(-decimal)

but that was not how they behaved in practice.

Due to the change in implementation, some very delicate tests may fail
that did not fail before. No extra failures were noted in scipy.

Closes #5200.