ENH: Improve announce to find github squash-merge commits.
Currently when the squash-merge button is hit on github, the PR title
with "(#PR)" appended is used as the commit message summary line when
the merge is fast-forwarded. This PR adds the ability to pick out the
corresponding PR numbers from the git log output.

Note that github may change this in the future and one early squash
merge in the 1.12.0 release did not have the PR number appended.