ENH: Refactor numpy ** operators for numpy scalar integer powers.
This change is the upshot of discussion on the numpy mailing list where
it was decided that a numpy scalar integer raised to a negative integer
power should raise a ValueError, as do Numpy arrays. In particular, we
want numpy scalars and numpy 0-d arrays to have similar behavior.
Previously zero was returned for this case.