BUG assert_almost_equal fails on subclasses that cannot handle bool
gh-8410 breaks a large number of astropy tests, because it sets up
a boolean array for values that should actually be compared (i.e.,
are not `nan` or `inf`) using `zeros_like`. The latter means that
for subclasses, the boolean test array is not a plain `ndarray` but
the subclass. But for astropy's `Quantity`, the `all` method is
undefined.

This commit ensures the test arrays from `isinf` and `isnan` are
used directly.