MAINT: Warn users when calling np.ma.MaskedArray.(arg-)partition function.
Using the np.median function on MaskedArrays uses the not-overriden
partition method of a plain np.ndarray without error or warning. (#7330)
This PR overrides the partition method on MaskedArrays but simply to
throw a Warning. This will make users aware that something ignores the
mask without breaking backwards-compatibility. This also applies to
the argpartition method (even if it's not called by np.median).