ENH: Implement axes keyword argument for gufuncs.
The axes argument allows one to specify the axes on which
the gufunc will operate (by default, the trailing ones).
It has to be a list with length equal to the number of
operands, and each element a tuple of length equal to the
number of core dimensions, with each element an axis index.

If there is only one core dimension, the tuple can be
replaced by a single index, and if none of the outputs have
core dimensions, the corresponding empty tuples can be
omitted.