BUG: Ensure Errors are correctly checked when PyFloat_AsDouble is called.
There was an error in np.random.uniform where if np.random.uniform
were called with a type that throwed exceptions when it was converted
to a float this exception wouldn't be raised.

This bug was due to an issue where PyFloat_AsDouble was called but
no check for PyErr_Occurred was performed after.

This PR fixes the issue by ensuring that Cython will always emit a
call to PyErr_Occurred if PyFloat_AsDouble returns -1.0

Fixes: #8865