BUG: Calling ufuncs with a positional output argument causes the result to inherit the output's mask
This brings `np.add(a, b, out)` in line with  `np.add(a, b, out=out)`.
These previously differed because gh-10459 causes them to call __array_wrap__ in different ways (with and without the output argument in the context tuple, respectively).

Since the data in the `out` argument is never used by ufuncs, it seems consistent that the mask should not be either.