MAINT: Remove use of unittest in NumPy tests.
This removes a few left over uses of unittest. The main changes apart
from removal of Test case are:

* `setUp` replaced by nose and pytest compatible `setup`
* `tearDown` replaced by nose and pytest compatible `teardown`
* `assertRaises` replaced by `assert_raises`
* `assertEqual` replaced by `assert_equal`

The last two are in `numpy/testings/tests/test_utils.py`, so may seem a
but circular, but at least are limited to those two functions.

The use of `setup` and `teardown`, can be fixed up with the pytest
equivalents after we have switched to pytest.